+++ server/tm_cmds.go
@@ -3,11 +3,14 @@ package server
 // DONTCOVER
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/spf13/cobra"
-	"github.com/tendermint/tendermint/p2p"
+	cfg "github.com/tendermint/tendermint/config"
 	pvm "github.com/tendermint/tendermint/privval"
+	"github.com/tendermint/tendermint/scripts/keymigrate"
+	"github.com/tendermint/tendermint/scripts/scmigrate"
 	tversion "github.com/tendermint/tendermint/version"
 	yaml "gopkg.in/yaml.v2"
 
@@ -113,4 +116,68 @@ against which this app has been compiled.
 			return nil
 		},
 	}
+}
+
+// makeKeyMigrateCmd is ported from tendermint's key-migrate command, but
+// uses the SDK's own server.Context.
+// ref: https://github.com/tendermint/tendermint/blob/master/UPGRADING.md#database-key-format-changes
+func makeKeyMigrateCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "key-migrate",
+		Short: "Run Tendermint database key migration",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			ctx, cancel := context.WithCancel(cmd.Context())
+			defer cancel()
+
+			serverCtx := GetServerContextFromCmd(cmd)
+			config := serverCtx.Config
+
+			contexts := []string{
+				// this is ordered to put the
+				// (presumably) biggest/most important
+				// subsets first.
+				"blockstore",
+				"state",
+				"peerstore",
+				"tx_index",
+				"evidence",
+				"light",
+			}
+
+			for idx, dbctx := range contexts {
+				serverCtx.Logger.Info("beginning a key migration",
+					"dbctx", dbctx,
+					"num", idx+1,
+					"total", len(contexts),
+				)
+
+				db, err := cfg.DefaultDBProvider(&cfg.DBContext{
+					ID:     dbctx,
+					Config: config,
+				})
+
+				if err != nil {
+					return fmt.Errorf("constructing database handle: %w", err)
+				}
+
+				if err = keymigrate.Migrate(ctx, db); err != nil {
+					return fmt.Errorf("running migration for context %q: %w",
+						dbctx, err)
+				}
+
+				if dbctx == "blockstore" {
+					if err := scmigrate.Migrate(ctx, db); err != nil {
+						return fmt.Errorf("running seen commit migration: %w", err)
+
+					}
+				}
+			}
+
+			serverCtx.Logger.Info("completed database migration successfully")
+
+			return nil
+		},
+	}
+
+	return cmd
 }